requirements.txt
Django==3.1
Pillow==7.1.1
django-crispy-forms==1.10.0
whitenoise==5.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6